# Improve performance of tri products for `PDiagMat`